requirements.txt
google-auth==1.6.3
requests==2.22.0